requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
mediapipe>=0.8.9
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
moviepy>=1.0.3
tqdm>=4.62.0
pyyaml>=5.4.1
pytest>=6.2.5
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
click>=8.0.0
beautifulsoup4>=4.9.3
requests>=2.26.0
PyPDF2>=3.0.0 